Put back sources on /
