+++ request.go
@@ -1,0 +1,7 @@
+package main
+
+// Request abstracts a request.
+type Request interface {
+	item() (item, error)
+	userCredentials() (UserCredentials, error)
+}
