--- sources/encryptionService.go
@@ -1,5 +1,0 @@
-package main
-
-type EncryptionService interface {
-	comparePasswords(ciphered string, plain string) bool
-}
