--- sources/main.go
@@ -1,11 +1,0 @@
-package main
-
-func main() {
-	app := App{ 
-				database: NewPostGreSQLDatabase(),
-				router: NewMuxRouter(),
-				encryptionService: NewBCryptEncryptionService(),
-				authService: NewJWTAuthService(),
-			}
-	app.run()
-}
